+++ CHANGELOG.md
@@ -32,6 +32,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     an option for extracting and formating the entire date. See
     docs/workflow/workflow-configuration-how-to.md for more information.
 
+### Changed
+
+- **CUMULUS-2725**
+  - Updated providers endpoint to return encrypted password
+  - Update providers model to try decrypting credentials before encryption to allow for better handling of updating providers
 
 ## [v9.8.0] 2021-10-19